fix(export): allow OCI worker to unlazy image output
To allow images to be exported all layers need to be unlazied.
Unlazy means to actually make sure that all layers exist.
Typically, this can mean that base images may not be downloaded.

This fixes an oversight in the logic.  It turns out that this
piece of code handles both create an image in an "image store"
as well as unlazy all layers for the content store.

The OCI worker does not have an image store.  Thus, the logic
of this function accidently excluded OCI workers from unlazy
of their layers for the content store.

Signed-off-by: Chris Goller <goller@gmail.com>