refactor(load): return manifests via solve response
Previously, manifests and configs were sent as annotations
on the image index. This caused failures when pushing to GCR
and oci-mediatypes was false.

Now a new export image option, `depot.export.image.version` has
been added.  If the version is set to `2` then we send back manifest
and config in the solve response.  If not set or `1` we'll continue
to send via annotations for backwards compatibility.

The response manifests and configs is a base64-encoded JSON array stored
at the key `depot.export.image.version`.

Signed-off-by: Chris Goller <goller@gmail.com>