feat: add creator and stable digests to disk usage
We want to know the original build and stable digests
associated with each snapshot and blob.

The goal was to have a minimally invasive patch.
That's a fancy way of saying I used context to propagate
the creator vertex and the stable digests into the metadata
storage.

The creator vertex will only ever be set when the data is
first created.

The stable digests are created within each `Op`. However,
those ops are written in such a way that they may not have
access to the digest later in their execution.

Signed-off-by: Chris Goller <goller@gmail.com>