perf: switch num concurrent layer fetch/push to 32
We are seeing slow container push and are wondering if
the global pool of concurrent layer pushes is causing
a slowdown.

This PR addresses both concurrent fetch and pulls as the
same global variable controls the semaphore weight.

Additionally, this adds printing of each layer as it is pushed
to a remote registry.

Signed-off-by: Chris Goller <goller@gmail.com>